Object Storage: NCP -> OCI로 변경
