+++ booquest-api/build.gradle
@@ -65,7 +65,7 @@ dependencies {
 	//postgresql
 	runtimeOnly 'org.postgresql:postgresql'
 
-	// NCP Object Storage
+	// OCI Object Storage (S3 compatible)
 	implementation 'com.amazonaws:aws-java-sdk-s3:1.12.565'
 
 	//test
